Convert the html files to php and include files
This commit will convert the html files to php and add an 'include'
folder with header and footer files included.

And will also rename the old index.html file to index_old.html file.

In addition, the Register icon has been converted to a link and has been
pointed to register.php